backend/task1/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
selenium==4.15.2
beautifulsoup4==4.12.2
google-generativeai==0.3.2
lxml==4.9.3
webdriver-manager==4.0.1
python-dotenv==1.0.0

